Add scene to debug aspect ratio issues
This scene can play a level inside a subviewport, and show
semi-transparent red bars when the viewport is too tall or too wide. So
bughunters of ratio issues know where is the limit of their search, by
ignoring anything that is far away in the red area.

Add Fray's End to the subviewport, but this is supposed to be replaced
by the scene to debug.

Note that playing a level through a subviewport has limitations. For
instance I wasn't able to use the story book and select a quest in
Fray's End.

The minimum and maximum aspect ratios in this commit are rather
arbitrary.